conversations: extract MessageList — completes the ConversationsPanel split (#1087)
Third and final piece of the ConversationsPanel split (after Composer + DraftCards).
Lift the transcript into conversations/MessageList.svelte: the `.messages` scroll
container, the message rendering (messageBlock + markdown-it), assistant
citations (citeState/handleCite), the streaming block, the ask-user card, and
the interleaved <DraftCards>. It owns scrollEl, citeState, pendingAnswerText, and
the auto-scroll effect, reaching the conversations/editor stores via getters, so
it only needs {tab, currentNotePath} as props.

The one cross-boundary concern — the context-rail "Create note" (#177) must read
a selection only if it lives in the transcript pane — is handled by an exported
`getPaneSelectionText()` the panel calls via bind:this, preserving exact behavior.

ConversationsPanel is now the panel shell: tab list, context rail (model/effort
pickers + Create note), resize, and the three children. 2041 → 510 LOC across the
whole #1087 effort. Composer 373 · MessageList 343 · DraftCards 735.

pnpm lint (0 errors) + full pnpm test (4043) green.
